requirements.txt
transformers>=4.54.0
torchmetrics>=0.7.0
scikit-learn>=1.7.0
opencv-python>=4.10.0
opencv-contrib-python>=4.10.0
scikit-image>=0.25.0
Pillow>=11.0.0
matplotlib>=3.10.0
imageio>=2.37.0
timm>=1.0.19
lpips>=0.1.4
easyocr>=1.7.2
accelerate>=1.9.0
spacy>=3.8.0
ftfy>=6.3.1
openai>=1.98.0
tabulate>=0.9.0
pyyaml>=6.0
shapely>=2.1.1
GPUtil>=1.4.0
psutil>=7.0.0
ultralytics
pytesseract>=0.3.13
RapidFuzz
pynvml
